feat: move concatenate_json_files and pull_netpune_data functions to json_utils.py
